A UPClass is a class the contains a resampling technique and computes the weights of the UP distirbution

UPClass is a class for the resampling technique and its use in the UP disitrbution paradigm.

# up     <- UPClass$new()

Format

An R6Class generator object

Details

It scale the data, initilize a resmapling technique and compute the weigths of the UP distirbution

Fields

Resampling:

a resampling technqiue (Default leva-one-out LOO)

x

: design points

y

: response

Xscaled

: Scaled design points

Yscaled

: Sacelde response

d

: dimension

n

: number of design points

rho:

weight parameters

variogramme:

weight paramter (default "gauss")

Methods

UPClass$new(x = NA, y = NA, Scale =TRUE, resampling_type = "LOO", ListIndices = NULL, kfold = 10, division_method ="random")

Creates a new UPClass object.

scaleinput(newData)

Scale new data.

unscaleoutput(values )

unscale scaled output

computeweights(scnewdata)

compute the up weights for a new set of data.

Examples

library(UP)
d               <- 2;
n               <- 16
X                   <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
y               <- apply(X, 1, branin)
aUP             <- UPClass$new(X,y,Scale =TRUE, resampling_type = "KFCV", kfold =5)