Universal Prediction distribution based Efficient Global Optimization UP-EGO

Global optimization using the Universal Prediction distribution

upego(model, fun, nsteps, lower, upper, seed = 1, parinit = NULL,
  control = NULL, EEIcontrol = NULL)

Arguments

model
the surrogate model

fun
the real function

nsteps
the number of points to be generated

lower
the lower bound of the design space

upper
the upper bound of the design space

seed
the random seed (default = 1)

parinit
inital points to be used in the optimization (default NULL)

control
the optimization control parameters (default NULL)

EEIcontrol
the optimization criterion parameters (default NULL)

Value

list of generated points and their values and the last updated surrogate model

Examples

#' library(UP)
d            <- 2;
n            <- 16
X            <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Y            <- apply(X, 1, branin)

upsm            <- UPSM$new(sm= krigingsm$new(), UP=UPClass$new(X,Y,Scale =TRUE) )
upego_res <-  upego( upsm,fun = branin, nsteps = 1, lower= c(0,0),upper = c(1,1) )

#> [1] "OPTIMIZATION step : 1 ..."
#> [1] "new point "
#>         x1    x2
#> [1,] 0.125 0.625
#> [1] "New value  8.43883174131785"
#> [1] "current best value :  2.94383349064161   inital :  2.94383349064161"

print( min(upego_res$last$get_DOE()$y ) )

#> [1] 2.943833