UP-SMART

Enhance the overall prediction capabilities of a surrogate model by the Universal Prediction distribution based Surrogate Modeling Adapative Refinement Strategy UP-SMART

upsmart(model, fun, nsteps, lower, upper, seed = 1, parinit = NULL,
  control = NULL, RefControl = NULL)

Arguments

model
the surrogate model

fun
the real function

nsteps
the number of points to be generated

lower
the lower bound of the design space

upper
the upper bound of the design space

seed
the random seed (default = 1)

parinit
inital points to be used in the optimization (default NULL)

control
the optimization control parameters (default NULL)

RefControl
the refienement criterion parameters (default NULL)

Value

list of generated points and their values and the last updated surrogate model

Examples

#' library(UP)
d            <- 2;
n            <- 16
X            <- expand.grid(x1=s <- seq(0,1, length=5), x2=s)
Xtest        <- expand.grid(x1=seq(0,1,length=5), x2=seq(0,1,length=4))
Y            <- apply(X, 1, branin)
sm           <- krigingsm$new()
sm$setDOE(X,Y)
sm$train()
upsmart_res  <- upsmart(sm,fun = branin,nsteps = 5, lower= c(0,0),upper = c(1,1))

#> [1] "UP-SMART iteration : 1"
#> [1] "UP-SMART iteration : 2"
#> [1] "UP-SMART iteration : 3"
#> [1] "UP-SMART iteration : 4"
#> [1] "UP-SMART iteration : 5"

print(upsmart_res$last$get_DOE())

#> $x
#>           x1        x2
#> 1  0.0000000 0.0000000
#> 2  0.2500000 0.0000000
#> 3  0.5000000 0.0000000
#> 4  0.7500000 0.0000000
#> 5  1.0000000 0.0000000
#> 6  0.0000000 0.2500000
#> 7  0.2500000 0.2500000
#> 8  0.5000000 0.2500000
#> 9  0.7500000 0.2500000
#> 10 1.0000000 0.2500000
#> 11 0.0000000 0.5000000
#> 12 0.2500000 0.5000000
#> 13 0.5000000 0.5000000
#> 14 0.7500000 0.5000000
#> 15 1.0000000 0.5000000
#> 16 0.0000000 0.7500000
#> 17 0.2500000 0.7500000
#> 18 0.5000000 0.7500000
#> 19 0.7500000 0.7500000
#> 20 1.0000000 0.7500000
#> 21 0.0000000 1.0000000
#> 22 0.2500000 1.0000000
#> 23 0.5000000 1.0000000
#> 24 0.7500000 1.0000000
#> 25 1.0000000 1.0000000
#> 26 0.8818768 0.3738098
#> 27 0.1122530 0.1039815
#> 28 0.3790677 1.0000000
#> 29 0.6236599 0.0000000
#> 30 0.6362611 1.0000000
#> 
#> $y
#>  [1] 305.956302  80.060129  10.218600  20.597105   9.503736 191.588546
#>  [7]  32.717656   3.185863  27.158559   2.943833 105.345789  13.500183
#> [13]  24.278127  61.845013  24.508931  47.228033  22.407710  73.495391
#> [19] 124.656467  74.199029  17.235277  59.440237 150.837655 215.592921
#> [25] 152.014126  23.531113 123.998113 118.075401   8.803617 193.809155
#>